+++ previous-versions/files/RELEASES.md
@@ -279,3 +279,13 @@ Based on:
 - Speakeasy CLI 1.53.0 (2.58.0) https://github.com/speakeasy-api/speakeasy
 ### Releases
 - [Go v0.22.0] https://github.com/codatio/client-sdk-go/releases/tag/files/v0.22.0 - files
+
+## 2023-09-01 15:57:40
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Files.yaml
+- Speakeasy CLI 1.77.1 (2.91.4) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [go v0.23.0] previous-versions/files
+### Releases
+- [Go v0.23.0] https://github.com/codatio/client-sdk-go/releases/tag/previous-versions/files/v0.23.0 - previous-versions/files